* feature/hj_front2:
  보스 프론트 연결 끝~
  손님 마이페이지 버튼 정렬, 마이페이지 글씨 개행, 전 페이지 헤더 통일, 세탁 요청 페이지 중복된 UI 개선, 손님 마이페이지 게시글 UI 개선
  충돌 수정
  Revert "충돌 해결을 위한 임시 커밋"
  충돌 해결을 위한 임시 커밋

# Conflicts:
#	app.js
#	config/config.json
#	controller/laundries.controller.js
#	controller/reviews.controller.js
#	repository/reviews.repository.js
#	service/reviews.service.js
#	views/boss.ejs
#	views/header.html
#	views/login.ejs
#	views/mypage.ejs
#	views/postlaundry.ejs
#	views/signup.ejs
#	views/userreviews.ejs

+++ service/reviews.service.js
@@ -4,13 +4,14 @@ class ReviewService {
   reviewRepository = new ReviewRepository();
 
   //리뷰 생성
-  createReview = async (id, grade, comment, user_id, laundry_id) => {
+  createReview = async (grade, comment, user_id, laundry_id, boss_id) => {
     const createReviewData = await this.reviewRepository.createReview(
       id,
       grade,
       comment,
       user_id,
-      laundry_id
+      laundry_id,
+      boss_id
     );
 
     return createReviewData;
@@ -27,6 +28,10 @@ class ReviewService {
       if (!findReview) throw new Error("리뷰가 존재하지 않아요.");
       if (User !== findReview.id) {
         return { errorMessage: error.message };
+      }
+
+      if (User !== findReview.id) {
+        return { errormessage: "작성자가 아닙니다." };
       }
 
       const updateReview = await this.reviewRepository.updateReview(
@@ -66,9 +71,13 @@ class ReviewService {
   };
 
   //리뷰 삭제
-  deleteReview = async (id) => {
+  deleteReview = async (id, User) => {
     const findReview = await this.reviewRepository.findReviewById(id);
     if (!findReview) throw new Error("리뷰가 존재하지 않아요.");
+
+    if (User !== findReview.id) {
+      return { errormessage: "작성자가 아닙니다." };
+    }
 
     await this.reviewRepository.deleteReview(id);
   };
